refactor: fetchData, jsonConfig util로 이동
